% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompress.R
\name{decompress}
\alias{decompress}
\title{Decompress a file or connection}
\usage{
decompress(file, dest = sub("\\\\.(gz|bz2|xz)$", "", file), force = FALSE)
}
\arguments{
\item{file}{Path to a compressed file, or a connection. See Details.}

\item{dest}{Path to the destination file, or a connection. Must be given
explicitly if \code{file} is not a path that exists.}

\item{force}{Set to \code{TRUE} to overwrite \code{file} with \code{dest}, if applicable.}
}
\value{
The destination \code{dest}, invisibly.
}
\description{
Decompress the full contents of a file (or connection) into another file (or
connection). This can be useful if you need to access a compressed file in
some external software that doesn't support decompression, or for
compatibility with a function that doesn't accept connections and doesn't
handle decompression.
}
\details{
If \code{file} is a connection, it will be wrapped with \code{\link[=gzcon]{gzcon()}}, automatically
opened if necessary, and is always closed (and destroyed) afterwards.

See the underlying \code{\link[=gzfile]{gzfile()}} for an authoritative list of supported
compression types. At time of writing at least \code{gzip}, \code{bzip2}, \code{xz} and
\code{lzma} are supported.

The batch size used for I/O operations can be configured by setting the
\code{mire.decompress.batch_size} option (in bytes). The default batch size is 50 MiB.

The emitted messages can be suppressed by setting the \code{mire.decompress.verbose}
option to \code{FALSE}.
}
\examples{
decompress(url("https://www.stats.ox.ac.uk/pub/datasets/csb/ch12.dat.gz"), "ch12.dat")
head(read.table("ch12.dat"))
unlink("ch12.dat") # Tidy up
}
\seealso{
\code{\link[=gzfile]{gzfile()}} and \code{\link[=gzcon]{gzcon()}} to read compressed files directly.
}
