% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_counts.R
\name{denoise_counts}
\alias{denoise_counts}
\title{Denoise and Quantify Taxa from Kraken2 Results}
\usage{
denoise_counts(
  kreports,
  kmers,
  umis,
  samples = NULL,
  taxon = c("d__Bacteria", "d__Fungi", "d__Viruses"),
  drop_unmatched_taxa = TRUE,
  quantile = 0.95,
  alpha = 0.05,
  min_reads = 3L,
  min_minimizer_n_unique = 3L,
  min_samples = 3L,
  min_kmer_len = 3L,
  min_cells = 3L,
  cor_threshold = 0,
  p_threshold = 0.05,
  padj_threshold = 0.05,
  ...,
  study = "current study",
  alternative = "greater",
  p.adjust = "BH",
  method = "spearman"
)
}
\arguments{
\item{kreports}{A character vector of paths to Kraken2 report files, one per sample.}

\item{kmers}{A character vector of paths to Parquet k-mer count tables, one per sample.}

\item{umis}{A character vector of paths to Parquet UMI count tables, one per sample.}

\item{samples}{A character of sample identifier for each element in \code{umis}.
If not provided, the names of the \code{umis} vector will be used.}

\item{drop_unmatched_taxa}{A boolean value, indicates whether taxa not found
in celllines data should be dropped. Default: \code{TRUE}.}

\item{quantile}{Probabilities with values in \verb{[0, 1]} specifying the quantile
to calculate.}

\item{alpha}{Level of significance.}

\item{min_reads}{An integer, the minimal number of the total reads to filter
taxa.}

\item{min_minimizer_n_unique}{An integer, the minimal number of the unique
number of minimizer to filter taxa.}

\item{min_samples}{An integer, the minimal number of samples per taxid.}

\item{min_kmer_len}{An integer, the minimal number of kmer to filter taxa.}

\item{min_cells}{An integer, the minimal number of cell per taxid.}

\item{cor_threshold}{Minimum correlation coefficient required in sample-level
filtering. Default is \code{0}.}

\item{p_threshold}{Significance threshold for sample-level correlation
p-values. Default is \code{0.05}.}

\item{padj_threshold}{Adjusted p-value threshold for barcode-level denoising.
Default is \code{0.05}.}

\item{...}{Other arguments passed to \link[stats:cor.test]{cor.test}.}

\item{study}{A string of the study name, used to differentiate with cell line
data.}

\item{alternative}{A string specifying the alternative hypothesis, must be
one of "two.sided", "greater" (default) or "less". You can specify just the
initial letter.}

\item{p.adjust}{Pvalue correction method, a character string. Can be
abbreviated. Details see \link[stats:p.adjust]{p.adjust}.}

\item{method}{A character string indicating which correlation coefficient is
to be used for the test. One of "pearson", "kendall", or "spearman", can be
abbreviated.}
}
\description{
This function integrates three signal denoising strategies—cell-line-based
contaminant filtering, sample-level correlation filtering, and barcode-level
correlation testing— to identify reliable microbial taxa and quantify their
abundance across samples. It returns a table of taxa counts (per barcode),
after filtering out likely contaminants and false positives.
}
