% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunzip.R
\name{gunzip}
\alias{gunzip}
\title{Decompress a file}
\usage{
gunzip(file, dest = sub("\\\\.(gz|bz2|xz)$", "", file), force = FALSE)
}
\arguments{
\item{file}{Path to a compressed file, or a connection. Connections will be
automatically opened if necessary, and are always closed (and destroyed)
afterwards.}

\item{dest}{Path to the destination file, or a connection. Must be given
explicitly if \code{file} is not a local file that exists.}

\item{force}{Set to \code{TRUE} to overwrite \code{file} with \code{dest}, if applicable.}
}
\value{
The destination \code{dest}, invisibly.
}
\description{
Decompress the full contents of a file (or connection) into another file (or
connection). This can be useful if you need to access a compressed file in
some external software that doesn't support decompression, or for
compatibility with a function that doesn't accept connections and doesn't
handle decompression.
}
\details{
See \code{\link[=gzfile]{gzfile()}} for supported compression types.

The batch size for I/O operations can be configured by setting the
\code{mire.gunzip.batch_size} option (in bytes). The default batch size is 50 MiB.
}
\examples{
gunzip(url("https://www.stats.ox.ac.uk/pub/datasets/csb/ch12.dat.gz"), "ch12.dat")
head(read.table("ch12.dat"))
unlink("ch12.dat") # Tidy up
}
\seealso{
\code{\link[=gzfile]{gzfile()}} and \code{\link[=gzcon]{gzcon()}} to read compressed files directly.
}
