% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/let.R
\name{let}
\alias{let}
\title{Evaluate an expression in a modified context}
\usage{
let(expr, ...)
}
\arguments{
\item{expr}{Expression to evaluate after modifications.}

\item{...}{Name-value pairs giving values for symbols in \code{expr}. Evaluated
sequentially, so you can refer to previous names in subsequent pairs.}
}
\value{
The value of evaluating \code{expr} in the modified context.
}
\description{
Similar to \code{\link[=with]{with()}}, but with an interface optimized for modifying single
symbols, sequential evaluation, and an argument order suitable for piping.
}
\examples{
let(f(1:5), f = mean)
let(mean(x), x = 1:5)

# Use previous assignments
let(x = 2, y = x + 3, y * y)
}
