% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsome.R
\name{mapsome}
\alias{mapsome}
\alias{\%>>\%}
\title{Apply a function over a potentially null value}
\usage{
x \%>>\% f

mapsome(x, f, ...)
}
\arguments{
\item{x}{An object.}

\item{f}{A function.}

\item{...}{Additional arguments passed to \code{f}.}
}
\value{
\code{NULL} if \code{x} is null. Otherwise the result of \code{f(x, ...)}.
}
\description{
Transform or process a potentially null value with a function that does not
handle \code{NULL} inputs gracefully. Can be used to propagate \code{NULL} values
through a pipeline of calls, or in combination with \verb{\%||\%} when optional
function arguments with a \code{NULL} default need processing if provided.
}
\details{
Inspired by the common monad known as Maybe or Option in various languages.
}
\examples{
try(strsplit(NULL, ""))
NULL \%>>\% \(x) strsplit(x, "")
"42" \%>>\% \(x) strsplit(x, "")

# The pipe-friendly function form.
NULL |> mapsome(strsplit, "")
}
