% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matpow.R
\name{matpow}
\alias{matpow}
\alias{\%^\%}
\title{Integer powers of a square matrix}
\usage{
x \%^\% n

matpow(x, n)
}
\arguments{
\item{x}{A numeric matrix.}

\item{n}{An integer power.}
}
\value{
A matrix.
}
\description{
Raise a square matrix to an integer power using exponentiation by squaring.
A power of zero results in the identity matrix of matching dimensions.
Negative powers operate on the inverse.
}
\examples{
matrix(1:4, 2, 2) \%^\% 2
matrix(1:4, 2, 2) \%^\% 1
matrix(1:4, 2, 2) \%^\% 0

# Negative powers invert
matrix(1:4, 2, 2) \%^\% -1
}
\seealso{
\code{\link[=solve]{solve()}} to invert a matrix.
}
