% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mire-package.R
\docType{package}
\name{mire-package}
\alias{mire}
\alias{mire-package}
\title{mire: Microbiome Integrated Reconstruction and Estimation}
\description{
An integrated framework for microbiome reconstruction from sequencing data. It leverages tools like Kraken2 for taxonomic classification and combines cell barcodes, UMIs, and k-mer-based quantification to reconstruct microbial signals. Designed for both bulk and single-cell sequencing data, the package enables taxonomic and quantitative profiling of microbial communities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/mire}
  \item Report bugs at \url{https://github.com/WangLabCSU/mire/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
