% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{new_data_frame}
\alias{new_data_frame}
\title{Equip an object with the attributes to make it a data frame}
\usage{
new_data_frame(x, n)
}
\arguments{
\item{x}{An object, should be a list with elements of equal length.}

\item{n}{An integer number of rows; the shared length of elements in \code{x}.}
}
\value{
\code{x} with the necessary attributes set to make it a data frame.
}
\description{
No type checking or conversions are performed on the inputs, in order to
save precious microseconds when it matters.
}
\examples{
new_data_frame(list(a = 1:2, b = 3:4), 2L)
}
\seealso{
\code{\link[vctrs:new_data_frame]{vctrs::new_data_frame()}} for a further optimized constructor.

\code{\link[=data.frame]{data.frame()}} for a more convenient constructor.
}
