% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\alias{permute.data.frame}
\title{Permute the order of elements in an object}
\usage{
permute(x, ...)

\method{permute}{data.frame}(x, ...)
}
\arguments{
\item{x}{A vector or data frame.}

\item{...}{Additional arguments passed on to methods.}
}
\value{
The object \code{x} with elements permuted in a random order.
}
\description{
A generic function to generate a random permutation of an object. The default
method is equivalent to \code{sample(x)} without surprises for numeric vectors of
length 1. The \code{data.frame} method permutes the order of rows.
}
\examples{
permute(1:5)
permute(5) # Compare to sample(5)

# The data.frame method permutes rows
permute(iris) |> head()
}
\seealso{
\code{\link[=sample]{sample()}} for general purpose random sampling.

Other sampling helpers: 
\code{\link{resample}()}
}
\concept{sampling helpers}
