% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_DESCRIPTION.R
\name{read_DESCRIPTION}
\alias{read_DESCRIPTION}
\alias{read_DESCRIPTION_from_tar}
\alias{read_DESCRIPTION_from_zip}
\title{Read the DESCRIPTION file from a package archive}
\usage{
read_DESCRIPTION_from_tar(archive_path, package_name, tar = Sys.getenv("tar"))

read_DESCRIPTION_from_zip(archive_path, package_name)
}
\arguments{
\item{archive_path}{Path to the package archive.}

\item{package_name}{Name of the package in the archive.}

\item{tar}{The tar program to use. See \code{\link[=untar]{untar()}}.}
}
\value{
A character matrix with a column for each field as given by \code{\link[=read.dcf]{read.dcf()}}.
}
\description{
Read the DESCRIPTION file from a package archive
}
\examples{
\dontrun{
read_DESCRIPTION_from_tar("../mire_0.1.0.tar.gz", "mire")
}
}
