% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_if}
\alias{replace_if}
\title{Replace values satisfying a predicate}
\usage{
replace_if(x, f, values)
}
\arguments{
\item{x}{A vector.}

\item{f}{A function, called with \code{x}. The return value is used to select
elements in \code{x} to replace with \code{values}.}

\item{values}{Values assigned to elements of \code{x} satisfying \code{f}.}
}
\value{
The vector \code{x} with elements given by \code{f(x)} replaced with \code{values}.
}
\description{
Replace values satisfying a predicate
}
\examples{
replace_if(c(1, NA, 3), is.na, 42)
}
\seealso{
\code{\link[=replace]{replace()}} which this thinly wraps.
}
