% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\alias{resample.data.frame}
\title{Sample elements from an object with replacement}
\usage{
resample(x, size = length(x), prob = NULL, ...)

\method{resample}{data.frame}(x, size = nrow(x), prob = NULL, ...)
}
\arguments{
\item{x}{A vector or data frame.}

\item{size}{A positive integer. The number of elements to sample.}

\item{prob}{A numeric vector of sampling probability weights for each element
in \code{x}.}

\item{...}{Additional arguments passed on to methods.}
}
\value{
An object with \code{size} elements drawn from \code{x} with replacement.
}
\description{
A generic function to draw elements from an object with replacement. The
default method is equivalent to \code{sample(x, replace = TRUE)} without surprises
for numeric vectors of length 1. The \code{data.frame} method resamples rows.
}
\examples{
resample(1:5)
resample(5) # Compare to sample(5, replace = TRUE)

# The data.frame method samples rows
head(iris) |> resample()
}
\seealso{
\code{\link[=sample]{sample()}} for general purpose random sampling.

Other sampling helpers: 
\code{\link{permute}()}
}
\concept{sampling helpers}
