% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsahmi-package.R
\docType{package}
\name{scmire-package}
\alias{scmire}
\alias{scmire-package}
\title{scmire: Single-Cell Microbiome Integrated Reconstruction and Estimation}
\description{
scmire provides tools for processing single-cell microbiome sequencing data by filtering Kraken2 classification results, reconstructing read-level information, and quantifying k-mers and unique molecular identifiers (UMIs) across taxonomic lineages. It enables accurate profiling of microbial composition at single-cell resolution by integrating barcode and UMI information to resolve read identities and aggregate k-mer counts at multiple taxonomic ranks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yunuuuu/scmire}
  \item Report bugs at \url{https://github.com/Yunuuuu/scmire/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
