% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapez.R
\name{trapez}
\alias{trapez}
\title{Trapezoidal integration}
\usage{
trapez(x, y)
}
\arguments{
\item{x, y}{Numeric vectors giving the coordinates of the sample points.}
}
\value{
A numeric vector of length 1.
}
\description{
Find the area under a curve of sample points with the trapezoidal rule.
}
\details{
\code{x} and \code{y} are sorted by \code{x} if needed.
}
\examples{
\dontshow{if (getRversion() >= 4.3) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
trapez(c(0, 1, 2), c(1, 2, 0))

# Calculate AUC for each subject in the Indometh dataset.
Indometh |> by(~ Subject, with, trapez(time, conc))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=integrate]{integrate()}} for general-purpose numeric integration.
}
