% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch}
\alias{batch}
\title{Create a batch indicator for splitting a vector}
\usage{
batch(x, n = NULL, size = NULL, balance = !is.null(n))
}
\arguments{
\item{x}{A vector.}

\item{n}{An integer. The number of batches to create.}

\item{size}{An integer. The size of batches to create.}

\item{balance}{Logical. Should batch sizes be (approximately) balanced?}
}
\value{
An integer vector of \code{length(x)} identifying which batch each
element of \code{x} belongs to.
}
\description{
Create an indicator for splitting a vector into \code{n} batches, or batches of a
given \code{size}.
}
\examples{
batch(LETTERS, 5)
batch(LETTERS, size = 5)
batch(LETTERS, size = 5, balance = TRUE)
}
\seealso{
\code{\link[=split]{split()}} to split a vector into a list of batches.
}
