% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Find the greatest common divisor}
\usage{
gcd(x)
}
\arguments{
\item{x}{An integer vector.}
}
\value{
A positive-valued integer vector of length 1. The value is \code{1} if
the GCD of non-missing elements is 1. The value is \code{NA} if there are any
missing values \emph{and} the GCD of non-missing elements is greater than 1,
or if \code{x} has length 0.
}
\description{
Find the greatest common divisor of all values in an integer vector by
repeated application of the Euclidean algorithm.
}
\details{
\code{x} is coerced to an integer vector if possible.

Evaluation short-circuits if the greatest common divisor of non-missing
elements reaches 1.
}
\section{Why?}{


Whenever I need GCD, I always seem to need \code{Reduce(pracma::gcd, x)},
so I wanted an efficient version of that. Additionally, being limited to
integers avoids issues with floating point arithmetic, and allows
short-circuiting and special handling of missing values when it comes to
GCD of 1.
}

\examples{
gcd(c(48, 18, 33))
gcd(c(48, NA, 33))
gcd(c(48, NA, 31))
}
