% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Find the greatest common divisor}
\usage{
gcd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of length 1.
}
\description{
Find the greatest common divisor of all values in a numeric vector by
repeated application of the Euclidean algorithm over all elements.
}
\details{
\code{x} is coerced to a numeric vector if possible.
}
\examples{
gcd(c(48, 18, 39))
}
