% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mire-package.R
\docType{package}
\name{mire-package}
\alias{mire}
\alias{mire-package}
\title{mire: Mikko's Miscellaneous R Extensions}
\description{
Another miscellaneous collection of utilities that make the author's experience using R a little bit more pleasant.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikmart/mire}
  \item \url{https://mikmart.github.io/mire/}
  \item Report bugs at \url{https://github.com/mikmart/mire/issues}
}

}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@protonmail.com}

}
\keyword{internal}
