% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mire-package.R
\docType{package}
\name{mire-package}
\alias{mire}
\alias{mire-package}
\title{mire: Single-Cell Microbiome Integrated Reconstruction and Estimation}
\description{
An integrated framework for high-resolution microbiome analysis in single-cell sequencing. Leveraging Kraken2 for initial taxonomic classification, the package performs barcode-aware filtering and reconstructs read-level information by integrating cell barcodes and UMIs. It quantifies k-mers and aggregates taxon-specific signals across multiple taxonomic ranks. This enables accurate reconstruction and quantitative profiling of microbial communities at single-cell resolution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/mire}
  \item Report bugs at \url{https://github.com/WangLabCSU/mire/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
