% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfill.R
\name{unfill}
\alias{unfill}
\title{Replace repeated elements in a vector}
\usage{
unfill(x, value = NA)
}
\arguments{
\item{x}{A vector.}

\item{value}{A replacement value of length 1.}
}
\value{
The vector \code{x} with adjacent repeated elements replaced with \code{value}.
}
\description{
Replace repeated elements in a vector
}
\examples{
unfill(rep(1:3, 2:4))
unfill(rep(c("A", "B"), 3:4), "")
}
\seealso{
\code{\link[tidyr:fill]{tidyr::fill()}} and \code{\link[vctrs:vec_fill_missing]{vctrs::vec_fill_missing()}} for the reverse operation.

\code{\link[=rle]{rle()}} to find runs of repeated elements.
}
