% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfill.R
\name{unfill}
\alias{unfill}
\title{Replace adjacent repeated elements in a vector}
\usage{
unfill(x, values = NA)
}
\arguments{
\item{x}{A vector.}

\item{values}{A vector of replacement values. Standard recycling rules apply.}
}
\value{
The vector \code{x} with adjacent repeated elements replaced with \code{values}.
}
\description{
Replace adjacent repeated elements in a vector. This can be useful for
creating clearly visually separated grouped displays.
}
\examples{
unfill(rep(1:3, 2:4))
unfill(rep(c("A", "B"), 3:4), "")
}
\seealso{
\code{\link[=uniql]{uniql()}} to identify non-repeated elements.

\code{\link[tidyr:fill]{tidyr::fill()}} and \code{\link[vctrs:vec_fill_missing]{vctrs::vec_fill_missing()}} for the reverse operation.
}
