% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniq.R
\name{uniq}
\alias{uniq}
\title{Find adjacent repeated elements in a vector}
\usage{
uniq(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical vector of \code{length(x)}.
}
\description{
Similarly to the Unix command \code{uniq}, find elements in a vector that are
not the same as the preceding value.
}
\examples{
uniq(c(1, 1, 2, 1))
}
\seealso{
\code{\link[=unfill]{unfill()}} to replace adjacent repeated elements.

\code{\link[=duplicated]{duplicated()}} to find repeated elements regardless of adjacency.
}
